start app in paused state

+++ app/src/main/java/com/example/e_inkfitness/core/model/UserPrefs.kt
@@ -6,5 +6,5 @@ enum class Units {
 
 data class User(
     val units: Units,
-    val weightKG: Int,
+    val weightKG: Float,
 )
